+++ src/components/RichEditor/SlateProvider.tsx
@@ -1,18 +1,8 @@
 'use client';
-import { useState, useCallback, useMemo, useEffect, ReactNode } from 'react';
-import { ElementRenderer, CustomLeaf } from './SlateRenderer';
-
-import { createEditor, Descendant, CustomTypes } from 'slate';
+import { useState, useEffect, ReactNode } from 'react';
+import { createEditor, Descendant } from 'slate';
 import { withHistory } from 'slate-history';
-import {
-  Slate,
-  Editable,
-  withReact,
-  useSlate,
-  RenderLeafProps,
-} from 'slate-react';
-import Toolbar from './Toolbar';
-import { Paragraph } from './slate';
+import { Slate, withReact } from 'slate-react';
 
 export default function SlateProvider({ children }: { children: ReactNode }) {
   const [editor] = useState(() => withHistory(withReact(createEditor())));
