+++ src/components/RichEditor/Toolbar.tsx
@@ -1,5 +1,5 @@
 import { useSlate } from 'slate-react';
-import { Editor, Transforms, Element, Node, Text } from 'slate';
+
 import { BiHeading } from 'react-icons/bi';
 import {
   BiAlignLeft,
@@ -26,6 +26,7 @@ import { BsCardHeading } from 'react-icons/bs';
 import Button from 'react-bootstrap/Button';
 import ButtonGroup from 'react-bootstrap/ButtonGroup';
 import ButtonToolbar from 'react-bootstrap/ButtonToolbar';
+import { EditorCmd } from './Commands';
 
 export default function Toolbar() {
   const editor = useSlate();
@@ -35,9 +36,7 @@ export default function Toolbar() {
       <ButtonGroup>
         <Button
           variant={variant}
-          onClick={() =>
-            CustomEditor.toggleParagraphProp(editor, { header: 1 })
-          }
+          onClick={() => EditorCmd.toggleParagraphProp(editor, { header: 1 })}
         >
           <BsCardHeading />
         </Button>
@@ -46,25 +45,25 @@ export default function Toolbar() {
       <ButtonGroup>
         <Button
           variant={variant}
-          onClick={() => CustomEditor.toggleMark(editor, 'bold')}
+          onClick={() => EditorCmd.toggleMark(editor, 'bold')}
         >
           <MdFormatBold />
         </Button>
         <Button
           variant={variant}
-          onClick={() => CustomEditor.toggleMark(editor, 'italic')}
+          onClick={() => EditorCmd.toggleMark(editor, 'italic')}
         >
           <MdFormatItalic />
         </Button>
         <Button
           variant={variant}
-          onClick={() => CustomEditor.toggleMark(editor, 'strike')}
+          onClick={() => EditorCmd.toggleMark(editor, 'strike')}
         >
           <MdFormatStrikethrough />
         </Button>
         <Button
           variant={variant}
-          onClick={() => CustomEditor.toggleMark(editor, 'underline')}
+          onClick={() => EditorCmd.toggleMark(editor, 'underline')}
         >
           <MdFormatUnderlined />
         </Button>
@@ -74,7 +73,7 @@ export default function Toolbar() {
         <Button
           variant={variant}
           onClick={() =>
-            CustomEditor.toggleParagraphProp(editor, { align: 'left' })
+            EditorCmd.toggleParagraphProp(editor, { align: 'left' })
           }
         >
           <BiAlignLeft />
@@ -82,7 +81,7 @@ export default function Toolbar() {
         <Button
           variant={variant}
           onClick={() =>
-            CustomEditor.toggleParagraphProp(editor, { align: 'center' })
+            EditorCmd.toggleParagraphProp(editor, { align: 'center' })
           }
         >
           <BiAlignMiddle />
@@ -90,7 +89,7 @@ export default function Toolbar() {
         <Button
           variant={variant}
           onClick={() =>
-            CustomEditor.toggleParagraphProp(editor, { align: 'right' })
+            EditorCmd.toggleParagraphProp(editor, { align: 'right' })
           }
         >
           <BiAlignRight />
@@ -98,7 +97,7 @@ export default function Toolbar() {
         <Button
           variant={variant}
           onClick={() =>
-            CustomEditor.toggleParagraphProp(editor, { align: 'justify' })
+            EditorCmd.toggleParagraphProp(editor, { align: 'justify' })
           }
         >
           <BiAlignJustify />
@@ -108,19 +107,19 @@ export default function Toolbar() {
       <ButtonGroup>
         <Button
           variant={variant}
-          onClick={() => CustomEditor.toggleBlock(editor, 'bullet-list')}
+          onClick={() => EditorCmd.toggleBlock(editor, 'bullet-list')}
         >
           <MdFormatListBulleted />
         </Button>
         <Button
           variant={variant}
-          onClick={() => CustomEditor.toggleBlock(editor, 'number-list')}
+          onClick={() => EditorCmd.toggleBlock(editor, 'number-list')}
         >
           <MdFormatListNumbered />
         </Button>
         <Button
           variant={variant}
-          onClick={() => CustomEditor.toggleBlock(editor, 'check-list')}
+          onClick={() => EditorCmd.toggleBlock(editor, 'check-list')}
         >
           <MdChecklist />
         </Button>
@@ -129,13 +128,13 @@ export default function Toolbar() {
       <ButtonGroup>
         <Button
           variant={variant}
-          onClick={() => CustomEditor.toggleBlock(editor, 'image')}
+          onClick={() => EditorCmd.toggleBlock(editor, 'image')}
         >
           <MdImage />
         </Button>
         <Button
           variant={variant}
-          onClick={() => CustomEditor.toggleBlock(editor, 'audio')}
+          onClick={() => EditorCmd.toggleBlock(editor, 'audio')}
         >
           <MdAudiotrack />
         </Button>
@@ -143,102 +142,3 @@ export default function Toolbar() {
     </ButtonToolbar>
   );
 }
-
-// helpers
-
-const LIST_TYPES = [
-  'number-list',
-  'bullet-list',
-  'check-list',
-] as Element['type'][];
-
-export const CustomEditor = {
-  toggleBlock: (editor: Editor, type: Element['type']) => {
-    const isList = LIST_TYPES.includes(type);
-    const isActive = CustomEditor.isActive(editor, type);
-
-    Transforms.unwrapNodes(editor, {
-      match: (n) =>
-        !Editor.isEditor(n) &&
-        Element.isElement(n) &&
-        LIST_TYPES.includes(n.type),
-      split: true,
-    });
-    let newProperties: Partial<Element> = {
-      type: isActive
-        ? 'paragraph'
-        : isList
-        ? type === 'check-list'
-          ? 'check-item'
-          : 'list-item'
-        : type,
-    };
-    Transforms.setNodes<Element>(editor, newProperties);
-    if (!isActive && isList) {
-      const block = { type: type, children: [] } as Element;
-      Transforms.wrapNodes(editor, block);
-    }
-  },
-
-  toggleParagraphProp: (
-    editor: Editor,
-    format: Pick<Paragraph, 'header' | 'align'>,
-  ) => {
-    format = CustomEditor.query<typeof format>(editor, (e) => {
-      if (!(e.type === 'paragraph')) return format;
-      if (e.align && e.align === format.align)
-        format = { ...format, align: undefined };
-      if (format.header)
-        format = {
-          ...format,
-          header: (e.header
-            ? e.header > 0
-              ? e.header - 1
-              : undefined
-            : 3) as (typeof format)['header'],
-        };
-      return format;
-    });
-
-    Transforms.setNodes<Paragraph>(editor, format, {
-      match: (n) => Element.isElement(n) && n.type === 'paragraph',
-    });
-  },
-
-  query<T>(editor: Editor, map: (e: Element) => T) {
-    const [result] = Array.from(
-      editor.nodes<Element>({
-        match: (n) => !Editor.isEditor(n) && Element.isElement(n) && !!map(n),
-      }),
-    );
-    // if (!result) return null;
-    return map(result[0]);
-  },
-
-  updateElement(editor: Editor, newProps: Partial<Element>) {
-    Transforms.setNodes<Element>(editor, newProps);
-  },
-
-  isActive: (editor: Editor, type: Element['type']) => {
-    const { selection } = editor;
-    if (!selection) return false;
-    const [match] = Array.from(
-      Editor.nodes(editor, {
-        at: Editor.unhangRange(editor, selection),
-        match: (n) =>
-          !Editor.isEditor(n) && Element.isElement(n) && n.type === type,
-      }),
-    );
-    return !!match;
-  },
-
-  isMarkActive: (editor: Editor, mark: keyof Omit<Text, 'text' | 'color'>) => {
-    const marks = Editor.marks(editor);
-    return (marks && marks[mark]) || false;
-  },
-  toggleMark: (editor: Editor, mark: keyof Omit<Text, 'text' | 'color'>) => {
-    const isActive = CustomEditor.isMarkActive(editor, mark);
-    if (isActive) Editor.removeMark(editor, mark);
-    else Editor.addMark(editor, mark, true);
-  },
-};
